fixed clippy and rust size
